move facility edit properties to own tab
