+++ client/packages/common/src/intl/locales/en/common.json
@@ -1059,6 +1059,7 @@
   "label.programs": "Programs",
   "label.programs-immunisations": "Immunizations",
   "label.projected": "Projected",
+  "label.properties": "Properties",
   "label.provider-name": "Provider name",
   "label.purchase-cost-price": "Purchase cost",
   "label.qty-item-name": "Qty/Item Name",
